defined the main structure to make queries and display the results.
If applied, the commit will have defined the main structure to make queries and display the results.

The connection is finally happening but the wrong query is being made or it's being displayed in the wrong way.

Need to define better the querie/response to have the desire output.

After this define the input filter usage.