Finally possible to ready correct data response from Graphql, this time using Axios. As well, changed the data from App.vue to MovieComponent.vue fot a better organization.
If applied, this commit will have finally displayed the correct data from Graphql query. For this was no longer used the setup() function and instead the mounted () with axios that fetches all the data and deposite it in movies state.

Need to figure out now how to re-fetch the same data through a input search.

All this state and actions should be transferred to the Vuex store.